Renamed main package to server
