+++ package.json
@@ -3,7 +3,7 @@
     "private": true,
     "version": "1.0.0",
     "description": "Arbitration Outsourcing for State Channels",
-    "main": "packages/main/lib/startUp.js",
+    "main": "packages/server/lib/startUp.js",
     "scripts": {
         "bootstrap": "lerna bootstrap --hoist",
         "bootstrap-ci": "lerna bootstrap --hoist --ci -- --only=prod",
