+++ packages/block/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pisa/block",
+  "name": "@pisa-research/block",
   "version": "0.1.7",
   "description": "Block utilites used in the PISA project",
   "author": "PISA Research Ltd",
