+++ packages/client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pisa/client",
+  "name": "@pisa-research/client",
   "version": "0.1.7",
   "description": "A client library for the PISA API.",
   "main": "lib/index.js",
