+++ packages/contracts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pisa/contracts",
+  "name": "@pisa-research/contracts",
   "version": "0.1.7",
   "description": "PISA contracts",
   "author": "PISA Research Ltd",
