+++ packages/server/package.json
@@ -1,8 +1,8 @@
 {
-    "name": "@pisa/main",
+    "name": "@pisa-research/server",
     "version": "0.1.7",
     "description": "Arbitration Outsourcing for State Channels",
-    "homepage": "https://github.com/PISAresearch/pisa/tree/master/packages/main#readme",
+    "homepage": "https://github.com/PISAresearch/pisa/tree/master/packages/server#readme",
     "license": "ISC",
     "main": "lib/startUp.js",
     "directories": {
