+++ packages/server/src/responder/component.ts
