+++ packages/server/src/service/favicon.ico
