+++ packages/server/src/startUp.ts
