+++ packages/utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pisa/utils",
+  "name": "@pisa-research/utils",
   "version": "0.1.7",
   "description": "Misc utilities used in the PISA project",
   "author": "PISA Research Ltd",
