+++ test/endtoend/serviceEndToEnd.test.ts
@@ -4,13 +4,13 @@ import chaiAsPromised from "chai-as-promised";
 import { KitsuneTools } from "../external/kitsune/tools";
 import { ethers } from "ethers";
 import { arrayify, BigNumber } from "ethers/utils";
-import { PisaService } from "../../packages/main/src/service/service";
-import config from "../../packages/main/src/service/config";
+import { PisaService } from "../../packages/server/src/service/service";
+import config from "../../packages/server/src/service/config";
 import Ganache from "ganache-core";
 import levelup, { LevelUp } from "levelup";
 import MemDown from "memdown";
 import encodingDown from "encoding-down";
-import { deployPisa } from "../../packages/main/__tests__/utils/contract";
+import { deployPisa } from "../../packages/server/__tests__/utils/contract";
 import { wait, expectAsync } from "../../packages/test-utils/src";
 import PisaClient from "../../packages/client";
 chai.use(chaiAsPromised);
